Fix Pigeon compatibility issues and improve message type safety

+++ flutter_custom_tabs_android/android/src/test/kotlin/com/github/droibit/flutter/plugins/customtabs/core/options/CustomTabsCloseButtonBuilderTest.kt
@@ -9,7 +9,7 @@ class CustomTabsCloseButtonBuilderTest {
     fun setOptions_withAllOptions() {
         val options = mapOf(
             "icon" to "ic_arrow_back",
-            "position" to CustomTabsIntent.CLOSE_BUTTON_POSITION_START
+            "position" to CustomTabsIntent.CLOSE_BUTTON_POSITION_START.toLong(),
         )
 
         val closeButton = CustomTabsCloseButton.Builder()
